[ttnctl-user] Moved login functionality to util
